+++ crates/subspace-runtime/Cargo.toml
@@ -46,6 +46,7 @@ sp-std = { version = "4.0.0-dev", default-features = false, git = "https://githu
 sp-transaction-pool = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate", rev = "26d69bcbe26f6b463e9374e1b1c54c3067fb6131" }
 sp-version = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate", rev = "26d69bcbe26f6b463e9374e1b1c54c3067fb6131" }
 subspace-core-primitives = { version = "0.1.0", default-features = false, path = "../subspace-core-primitives" }
+subspace-primitives = { version = "0.1.0", default-features = false, path = "../subspace-primitives" }
 
 # Used for the node template's RPCs
 frame-system-rpc-runtime-api = { version = "4.0.0-dev", default-features = false, git = "https://github.com/paritytech/substrate", rev = "26d69bcbe26f6b463e9374e1b1c54c3067fb6131" }
@@ -92,6 +93,7 @@ std = [
 	"sp-transaction-pool/std",
 	"sp-version/std",
 	"subspace-core-primitives/std",
+	"subspace-primitives/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
